fix(replays): moves replay list fetch error UI into table
